+++ tests/tests.go
@@ -24,76 +24,6 @@ type runner struct {
 	*testing.T
 }
 
-// WriteFile writes the byte contents to the given path, which should be
-// relative to this test's repository directory.
-func (r *runner) WriteFile(path string, contents []byte) {
-	err := ioutil.WriteFile(r.repoPath(path), contents, 0755)
-	if err != nil {
-		r.Fatal(err)
-	}
-}
-
-// ReadFile reads the byte contents of the given path, which should be relative
-// to this test's repository directory.
-func (r *runner) ReadFile(path string) []byte {
-	by, err := ioutil.ReadFile(r.repoPath(path))
-	if err != nil {
-		r.Fatal(err)
-	}
-	return by
-}
-
-func (r *runner) Cd(path string) {
-	fullPath := r.repoPath(path)
-	r.Logf("$ cd %s", fullPath)
-	if err := os.Chdir(fullPath); err != nil {
-		r.Fatal(err)
-	}
-}
-
-func (r *runner) MkdirP(path string) {
-	fullPath := r.repoPath(path)
-	r.Logf("$ mkdir -p %s", fullPath)
-	if err := os.MkdirAll(fullPath, 0755); err != nil {
-		r.Fatal(err)
-	}
-}
-
-// GitBlob gets the blob OID of the given path at the given commit.
-func (r *runner) GitBlob(commitish, path string) string {
-	out := r.Git("ls-tree", commitish)
-	lines := strings.Split(out, "\n")
-	for _, line := range lines {
-		tabs := strings.Split(line, "\t")
-		if len(tabs) < 2 {
-			continue
-		}
-
-		attrs := strings.Split(tabs[0], " ")
-		if len(attrs) < 3 {
-			continue
-		}
-
-		if tabs[1] == path {
-			return attrs[2]
-		}
-	}
-
-	return ""
-}
-
-// Git executes a Git command and returns the combined STDOUT and STDERR.
-func (r *runner) Git(args ...string) string {
-	name := "git"
-	if args != nil && len(args) > 0 && args[0] == "lfs" {
-		name = bin
-		args = args[1:len(args)]
-	}
-
-	cmd := exec.Command(name, args...)
-	return r.execCmd(cmd)
-}
-
 func (r *runner) execCmd(cmd *exec.Cmd) string {
 	r.logCmd(cmd.Args[0], cmd.Args[1:len(cmd.Args)]...)
